+++ src/zepben/eas/client/work_package.py
@@ -43,6 +43,7 @@ __all__ = [
     "ForecastConfig",
     "FeederConfig",
     "FeederConfigs",
+    "NodeLevelResultsConfig"
 ]
 
 
@@ -330,8 +331,23 @@ class ModelConfig:
     Note the LV line fixer will fix all conductors upstream of the consumer up to the distribution transformer they are connected to.
     """
 
+    simplify_network: Optional[bool] = None
+    """
+    Flag to control whether to simplify the network model before translation.
+    """
+
     collapse_lv_networks: Optional[bool] = None
     """Flag to control whether to collapse lv network in the model."""
+
+    collapse_negligible_impedances: Optional[bool] = None
+    """
+    Flag to control whether to collapse conductors with negligible impedance during network simplification.
+    """
+
+    combine_common_impedances: Optional[bool] = None
+    """
+    Flag to control whether to combine conductors with common impedance during network simplification.
+    """
 
     feeder_scenario_allocation_strategy: Optional[FeederScenarioAllocationStrategy] = None
     """
@@ -557,6 +573,54 @@ class StoredResultsConfig:
 
 
 @dataclass
+class NodeLevelResultsConfig:
+    """
+    Configuration settings for node level results.
+    """
+
+    collect_voltage: Optional[bool] = None
+    """
+    Include voltage values in node level results.
+    """
+
+    collect_current: Optional[bool] = None
+    """
+    Include current values in node level results
+    """
+
+    collect_power: Optional[bool] = None
+    """
+    Include power values in node level results
+    """
+
+    mrids_to_collect: Optional[List[str]] = None
+    """
+    A list of MRID's to collect node level results at. Note: Depending on the network simplification
+    and translation these mrid's may not exist in the final OpenDss and no results will be collected.
+    """
+
+    collect_all_switches: Optional[bool] = None
+    """
+    Collect node level results at all switches in the network.
+    """
+
+    collect_all_transformers: Optional[bool] = None
+    """
+    Collect node level results at all transformers in the network.
+    """
+
+    collect_all_conductors: Optional[bool] = None
+    """
+    Collect node level results at all conductors in the network.
+    """
+
+    collect_all_energy_consumers: Optional[bool] = None
+    """
+    collect node level results at all energy consumers in the network.
+    """
+
+
+@dataclass
 class GeneratorConfig:
     """
     Configuration settings for the OpenDSS model.
@@ -566,6 +630,7 @@ class GeneratorConfig:
     model: Optional[ModelConfig] = None
     solve: Optional[SolveConfig] = None
     raw_results: Optional[RawResultsConfig] = None
+    node_level_results: Optional[NodeLevelResultsConfig] = None
 
 
 @dataclass