+++ changelog.md
@@ -17,11 +17,12 @@
 
 ## [0.22.0] - 2025-08-14
 ### Breaking Changes
-* None.
+* TODO: new generatorConfig fields.
 
 ### New Features
 * Added basic client method `run_ingestor` to run ingestors via EAS's `executeIngestor` graphql mutation. 
 * Added basic client methods `get_ingestor_run` and `get_ingestor_run_list` to retrieve the records of previous ingestor runs.
+* Support node level results
 
 ### Enhancements